+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+## Added
+- Attractor camera.
+- Keyboard C cycles through cameras.
 
 ## [0.8.0] - 2017-04-16
 ### Added